Implement ProductService with ProductServiceImpl class

+++ src/main/java/com/kyle/restws/ProductService.java
@@ -6,6 +6,6 @@ public interface ProductService {
 
 	List <Product> getProducts();
 	
-	int addProduct(Product product);
+	long addProduct(Product product);
 	
 }
